+++ Sources/SwiftTUIRuntime/RunLoop/RunLoop.swift
@@ -229,6 +229,7 @@ public final class RunLoop<State: Equatable & Sendable, Content: View> {
   package var pendingCoalescedEventBatches = 0
   package var pendingCoalescedWakeCauses: Set<WakeCause> = []
   package var cancelledRenderCount = 0
+  package var nextSemanticHostFrameSequence: UInt64 = 0
   package var deferredLifecycleCarryForward: [LifecycleCommitEntry] = []
   package var reportedRuntimeIssues: Set<RuntimeIssue> = []